requirements.txt
matplotlib.pyplot
pandas
requests
bs4
